Serve FAT32 on the read path instead of only detecting it
fat_parse_boot recognised FAT32 and then every FAT-table access failed
WASMOS_ERR_FS_CORRUPT, so the volume mounted unusable. Three separate things
had to change, and only the third makes anything work end to end.

1. CLUSTER WIDTH. Roughly twenty-five uint16_t cluster fields become uint32_t,
   across every scan context, the chain and chain-walk contexts, the open-file
   table, the FAT-entry context and fat_lba_for_cluster. This is the part with
   teeth: a missed field does not error, it truncates a cluster number and
   reads the wrong sector. test_fat32_start_cluster_above_16_bits_round_trips
   uses a start cluster of 0x12345 for exactly that reason -- while the field
   was 16-bit the compiler reported the case's own comparison as
   "always false", which is the truncation stated as a diagnostic.

2. FAT32 FAT ENTRIES. 28 significant bits, so reads mask and chain_next
   compares against 0x0FFFFFF8. The write path preserves the reserved top
   nibble rather than clearing it: those bits belong to the volume, not to the
   value being stored. A FAT32 entry is 4 bytes and, because bytes_per_sector
   is a multiple of 4, never straddles a sector edge -- unlike the FAT12 case
   the same function still handles.

   The start cluster also stops being a single 16-bit field: FAT splits it
   across dirent bytes 26..27 and 20..21. fat_dirent_cluster /
   fat_dirent_set_cluster are now the only places that know this. The high half
   is a reserved zero on FAT12/16, so both are correct on every volume type and
   a FAT12/16 entry stays byte-identical to what the old two-byte store wrote.

3. THE ROOT BECOMES A CHAIN. FAT32 has no fixed root region: no
   root_entry_count, no root_dir_lba. Six sites open-coded the fixed-root
   assumption (both resolvers at entry and at '..', readdir, chdir), and four of
   them additionally refused outright on root_entry_count == 0. They now go
   through one fat_root_origin(), which reports FAT12/16's fixed region as
   root=1 and FAT32's root as root=0 plus a start cluster -- so on FAT32 the
   root scans through the same chain-walking path as any subdirectory, which
   cfaa29ecac had just finished building. A FAT32 volume whose BPB_RootClus is
   unusable is now refused at mount rather than mounted and failing later.

VERIFICATION. The usual red-then-green was not available here: the width change
is required for the suite to COMPILE, so the FAT32 cases could not first fail
behaviourally. Mutation testing stands in, each mutation applied to the finished
tree and reverted:

  - drop the FAT32 branch in fat_fatent_read -> the chain and the
    above-16-bits cases fail
  - make fat_root_origin treat FAT32 as a fixed root -> the root lookup and
    root readdir cases fail
  - stop parsing BPB_RootClus -> the mount is refused and the suite aborts on
    its own precondition assert
  - zero the dirent high half -> only the above-16-bits case fails

Each mutation fails the cases that name the behaviour it removed and no others,
which is the property the red-first order normally establishes.

The FAT32 test volume is spec-legal rather than merely accepted. fat_parse_boot
keys FAT32 off root_entry_count == 0 and would have taken a tiny image, but a
cluster count below 65525 is a FAT16 volume by the specification, so a test
relying on that shortcut would not be testing FAT32. Hence ~69k clusters; calloc
leaves the 35 MB lazily mapped and the cases touch a handful of sectors.

NOT DONE, and now visible: mutation on FAT32 is untested, because the suite's
block layer refuses writes and create/mkdir/unlink all route through
fat_find_free_dir_slots, which is still single-cluster. Also filed: CHDIR does
not use fat_find_in_dir but carries its own hand-written scan, which still stops
at the first cluster. Now that every other read-side scan walks the chain that
is observable -- `ls /a/b` finds an entry `cd /a/b` reports as missing. Both are
in TASKS.md.

Gates: run-kernel-unit-tests green including the fourteen-case suite,
run-qemu-test green, run-qemu-cli-test 69/69, fmt-check and lint green. The
QEMU runs matter more than usual here: the boot volume is FAT16, so they are
the regression check on the cluster-width change.

+++ src/drivers/fs_fat/fat_alloc.h
@@ -14,11 +14,11 @@
 #include "fat_geom.h"
 #include "fat_types.h"
 
-/* Pure geometry helpers.  The end-of-chain marker is 0x0FFF on FAT12 and 0xFFFF
- * on FAT16; it is 0 for any other volume type, which callers treat as
- * unsupported rather than as a valid marker.  fat_total_clusters returns 0 when
- * the geometry is inconsistent. */
-uint16_t fat_end_of_chain_marker(const fat_mount_t* mnt);
+/* Pure geometry helpers.  The end-of-chain marker is 0x0FFF on FAT12, 0xFFFF on
+ * FAT16 and 0x0FFFFFFF on FAT32 (28 significant bits); it is 0 for an unknown
+ * volume type, which callers treat as unsupported rather than as a valid
+ * marker.  fat_total_clusters returns 0 when the geometry is inconsistent. */
+uint32_t fat_end_of_chain_marker(const fat_mount_t* mnt);
 uint32_t fat_total_clusters(const fat_mount_t* mnt);
 
 /* Read the FAT-table entry for e->cluster into e->value.  FAT_R_ERR on a
